ingestion-python/requirements.txt
-c constraints.txt
google-cloud-pubsub==0.41.0 \
    --hash=sha256:91e2d1ccef131224fb570bc72127de3a723a865cb0d5bac10f6db41e895d14d4 \
    --hash=sha256:ca8dc397f214419182f72c61bd90d7ba9cf41d2be41bc5df98d46f26a7e11506
google-cloud-storage==1.16.0 \
    --hash=sha256:aad501a7b47a92b20ce998208fc706959805982d639cc178f079382c43161283 \
    --hash=sha256:fe98ecfba927f8de626b315cffe68d7c4a8da3658f95fee382269b92e3d71328
grpcio==1.21.1 \
    --hash=sha256:0232add03144dd3cf9b660e2718244cb8e175370dca4d3855cb4e489a7811b53 \
    --hash=sha256:0f20e6dcb1b8662cdca033bb97c0a8116a5343e3ebc7f71c5fe7f89039978350 \
    --hash=sha256:10b07a623d33d4966f45c85d410bc6a79c5ac6341f06c3beda6c22be12cbfe07 \
    --hash=sha256:10c0476d5a52d21f402fc073745dc43b87cc8e080a1f49bbff4e1059019310fb \
    --hash=sha256:289dae0b35c59d191c524e976dd0a6f8c995d2062e72621eb866ad0f4472a635 \
    --hash=sha256:2be726f16142d358a0df1e81d583d6820ee561a7856a79cca2fbe49989308be7 \
    --hash=sha256:4338d2a81f5b4ca022e085040b3cfce19419a5ce44aa7e6810ac1df05365bed7 \
    --hash=sha256:4c535b46f20e66bee3097583231977e721acdfcb1671d1490c99b7be8902ce18 \
    --hash=sha256:557154aef70a0e979700cc9528bc8b606b668084a29a0d57dbc4b06b078a2f1c \
    --hash=sha256:5bfdd7e6647498f979dc46583723c852d97b25afe995d55aa1c76a5f9816bc1f \
    --hash=sha256:87d8943ae7aa6ca5bbad732867d7f17d2550e4966a0c15b52088e8b579422e47 \
    --hash=sha256:89d8719d8de4d137678f7caa979e1b0a6fd4026f8096ceef8c2d164bbabefaf2 \
    --hash=sha256:9c3f4af989ce860710ac1864dc2e867dd87e6cee51a2368df1b253596868e52f \
    --hash=sha256:9da52c3c728883aee429bb7c315049f50b2139f680cd86bb1165418e4f93a982 \
    --hash=sha256:9e9736659987beab42d18525ed10d21f80a1ba8389eac03425fbfd5684e6bbf0 \
    --hash=sha256:9ebcbb1a054cab362d29d3be571d43d6b9b23302d9fc4b43e5327000da1680a9 \
    --hash=sha256:a93e08636623e24c939851e2e0c0140b14f524b2980c9cdc4ea52b70a871c7e0 \
    --hash=sha256:ac322d86d1a079e0a118d544443ee16f320af0062c191b4754c0c6ec2fc79310 \
    --hash=sha256:b1fb101459868f52df6b61e7bb13375e50badf17a160e39fe1d51ae19e53f461 \
    --hash=sha256:b39aac96cceac624a23d540473835086a3ffa77c91030189988c073488434493 \
    --hash=sha256:b65507bc273c6dbf539175a786a344cc0ac78d50e5584f72c6599733f8a3301f \
    --hash=sha256:be5bb6e47417e537c884a2e2ff2e1a8b2c064a998fcfdfcc67528d4e63e7ebaf \
    --hash=sha256:c92de6a28a909c4f460dc1bbbcb50d676cf0b1f40224b222761f73fdd851b522 \
    --hash=sha256:c9f5962eb7fa7607b20eb0e4f59ed35829bd600fc0eacb626a6db83229a3e445 \
    --hash=sha256:d00bdf9c546ed6e649f785c55b05288e8b2dbb6bf2eb74b6c579fa0d591d35bd \
    --hash=sha256:da804b1dd8293bd9d61b1e6ea989c887ba042a808a4fbdd80001cfa059aafed2 \
    --hash=sha256:ead6c5aa3e807345913649c3be395aaca2bbb2d225f18b8f31f37eab225508f6 \
    --hash=sha256:eb4d81550ce6f826af4ec6e8d98be347fe96291d718bf115c3f254621ae8d98d \
    --hash=sha256:ef6a18ec8fd32ec81748fe720544ea2fb2d2dc50fd6d06739d5e2eb8f0626a1c \
    --hash=sha256:fad42835656e0b6d3b7ffc900598e776722e30f43b7234a48f2576ca30f31a47 \
    --hash=sha256:fb98dbfee0d963b49ae5754554028cf62e6bd695f22de16d242ba9d2f0b7339b \
    --hash=sha256:fb9cd9bb8d26dc17c2dd715a46bca3a879ec8283879b164e85863110dc6e3b2a
grpclib==0.3.0rc3 \
    --hash=sha256:617f31edbfa9f2279596926d28a69cb0443b69726f5fbc8ccb84df3d3201309c
pytest-black==0.3.5 \
    --hash=sha256:6849e2cfa9a659377c52e0e698e8d0f8a79ff603cb1c930ad86b512357edba4f
pytest-cov==2.7.1 \
    --hash=sha256:2b097cde81a302e1047331b48cadacf23577e431b61e9c6f49a1170bbe3d3da6 \
    --hash=sha256:e00ea4fdde970725482f1f35630d12f074e121a23801aabf2ae154ec6bdd343a
pytest-docstyle==1.5.0 \
    --hash=sha256:dcc54084b8e8282a83e50c6220c85d1c7d05e3871f74f0e911499b4f3adea756
pytest-env==0.6.2 \
    --hash=sha256:7e94956aef7f2764f3c147d216ce066bf6c42948bb9e293169b1b1c880a580c2
pytest-flake8==1.0.4 \
    --hash=sha256:4d225c13e787471502ff94409dcf6f7927049b2ec251c63b764a4b17447b60c0 \
    --hash=sha256:d7e2b6b274a255b7ae35e9224c85294b471a83b76ecb6bd53c337ae977a499af
pytest-mock==1.10.4 \
    --hash=sha256:43ce4e9dd5074993e7c021bb1c22cbb5363e612a2b5a76bc6d956775b10758b7 \
    --hash=sha256:5bf5771b1db93beac965a7347dc81c675ec4090cb841e49d9d34637a25c30568
pytest-mypy==0.3.2 \
    --hash=sha256:8f6436eed8118afd6c10a82b3b60fb537336736b0fd7a29262a656ac42ce01ac \
    --hash=sha256:acc653210e7d8d5c72845a5248f00fd33f4f3379ca13fe56cfc7b749b5655c3e
pytest-timeout==1.3.3 \
    --hash=sha256:4a30ba76837a32c7b7cd5c84ee9933fde4b9022b0cd20ea7d4a577c2a1649fb1 \
    --hash=sha256:d49f618c6448c14168773b6cdda022764c63ea80d42274e3156787e8088d04c6
pytest==4.5.0 \
    --hash=sha256:1a8aa4fa958f8f451ac5441f3ac130d9fc86ea38780dd2715e6d5c5882700b24 \
    --hash=sha256:b8bf138592384bd4e87338cb0f256bf5f615398a649d4bd83915f0e4047a5ca6
psutil==5.6.2 \
    --hash=sha256:206eb909aa8878101d0eca07f4b31889c748f34ed6820a12eb3168c7aa17478e \
    --hash=sha256:649f7ffc02114dced8fbd08afcd021af75f5f5b2311bc0e69e53e8f100fe296f \
    --hash=sha256:6ebf2b9c996bb8c7198b385bade468ac8068ad8b78c54a58ff288cd5f61992c7 \
    --hash=sha256:753c5988edc07da00dafd6d3d279d41f98c62cd4d3a548c4d05741a023b0c2e7 \
    --hash=sha256:76fb0956d6d50e68e3f22e7cc983acf4e243dc0fcc32fd693d398cb21c928802 \
    --hash=sha256:828e1c3ca6756c54ac00f1427fdac8b12e21b8a068c3bb9b631a1734cada25ed \
    --hash=sha256:a4c62319ec6bf2b3570487dd72d471307ae5495ce3802c1be81b8a22e438b4bc \
    --hash=sha256:acba1df9da3983ec3c9c963adaaf530fcb4be0cd400a8294f1ecc2db56499ddd \
    --hash=sha256:ef342cb7d9b60e6100364f50c57fa3a77d02ff8665d5b956746ac01901247ac4
ujson==1.35 \
    --hash=sha256:f66073e5506e91d204ab0c614a148d5aa938bdbf104751be66f8ad7a222f5f86
